requirements.txt

numpy>=1.21.0
tensorflow>=2.12,<2.16
opencv-python>=4.5.0
matplotlib>=3.4.0
python-dotenv>=1.0.0
scikit-image>=0.20.0
tqdm>=4.65.0
